refactor: onboarding steps
